fixed Makefile; improved  function
