+++ kutana/__init__.py
@@ -1,20 +1,15 @@
 """Kutana - engine for developing bots."""
 
-from kutana.logger import logger
-
-from kutana.kutana import Kutana
-from kutana.plugin import Plugin, Message, Attachment
-from kutana.executor import Executor
-from kutana.exceptions import ExitException
-
 from kutana.environment import Environment
-
+from kutana.exceptions import ExitException
+from kutana.executor import Executor
+from kutana.functions import get_path, import_module, load_plugins, load_value
+from kutana.kutana import Kutana
+from kutana.logger import logger
 from kutana.manager.basic import BasicManager
-from kutana.manager.debug import DebugManager, DebugEnvironment
-from kutana.manager.vk import VKManager, VKRequest, VKResponse, VKEnvironment
-from kutana.manager.tg import TGManager, TGResponse, TGEnvironment
-
-from kutana.functions import get_path, load_value, import_module, \
-    load_plugins, is_done
+from kutana.manager.debug import DebugEnvironment, DebugManager
+from kutana.manager.tg import TGEnvironment, TGManager, TGResponse
+from kutana.manager.vk import VKEnvironment, VKManager, VKRequest, VKResponse
+from kutana.plugin import Attachment, Message, Plugin
 
 NAME = "kutana"
